# server/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample server for a ticket purchasing application.",
    "version": "1.0.0",
    "title": "Ticket Purchasing API"
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "schemes": ["http"],
  "paths": {
    "/tickets": {
      "post": {
        "summary": "Create a new ticket",
        "description": "Creates a new ticket",
        "operationId": "createTicket",
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "ticket",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TicketReq"
            }
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "$ref": "#/definitions/Ticket"
            }
          },
          "400": {
            "description": "Invalid request data",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/tickets/{id}": {
      "get": {
        "summary": "Get ticket by ID",
        "description": "Returns a ticket by ID",
        "operationId": "getTicketById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Ticket"
            }
          },
          "404": {
            "description": "Ticket 1 not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/tickets/{id}/purchase": {
      "post": {
        "summary": "Purchase a ticket",
        "description": "Purchases a ticket",
        "operationId": "purchaseTicket",
        "consumes": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "purchase",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PurchaseReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ticket purchased successfully"
          },
          "400": {
            "description": "Invalid request data",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Ticket not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Ticket": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "example": 1
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "allocation": {
          "type": "integer",
          "format": "int32",
          "minimum": 1
        }
      }
    },
    "TicketReq": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "required": true
        },
        "description": {
          "type": "string"
        },
        "allocation": {
          "type": "integer",
          "format": "int32",
          "required": true,
          "minimum": 1,
          "example": 100
        }
      },
      "required": ["name", "allocation"]
    },
    "PurchaseReq": {
      "type": "object",
      "properties": {
        "quantity": {
          "type": "integer",
          "format": "int32",
          "required": true,
          "minimum": 1,
          "example": 2
        },
        "user_id": {
          "type": "integer",
          "description": "Dummy user ID for now",
          "format": "int64",
          "required": true,
          "example": 1
        }
      },
      "required": ["quantity", "user_id"]
    },
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "description": "Error message describing the issue"
        }
      },
      "required": ["error"]
    }
  }
}